+++ src/pages/AboutMePage/ui/ExplodingText/CodeSnippet/type.ts
@@ -1,0 +1,5 @@
+export type CodeSnippetProps = {
+	text: string;
+	position: [number, number, number];
+	onComplete?: () => void;
+};