+++ src/pages/AboutMePage/ui/ExplodingText/RotatingBgModel/type.ts
@@ -1,0 +1,7 @@
+export type RotatingBgModelProps = {
+	modelPath: string;
+	rotationDirection: number;
+	basePosition: [number, number, number];
+	shakeOffset: [number, number];
+	snippetZ?: number;
+};