+++ src/pages/AboutMePage/ui/ExplodingText/index.ts
@@ -1,0 +1,1 @@
+export { ExplodingText } from './ExplodingText';