+++ src/shared/lib/connectGlslForWebGl/type.ts
@@ -1,0 +1,7 @@
+import { RefObject } from 'react';
+
+export type ShaderRendererOptions = {
+	canvasRef: RefObject<HTMLCanvasElement>;
+	folder: string; // путь до папки glsl/`
+	timeSpeed?: number;
+};