+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.92.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.91.0...v1.92.0) (2025-05-16)
+
+### Features
+
+-   **models-lang:** added new model in about-page ([96d4d30](https://github.com/DKMFzF/frontend-portfolio/commit/96d4d3009adef5a8b0cab19d2fb6c5a6e9df0c74))
+
+## [1.91.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.90.0...v1.91.0) (2025-05-16)
+
+### Features
+
+-   **model-about-me:** added test model in page ([0c9ef7b](https://github.com/DKMFzF/frontend-portfolio/commit/0c9ef7b0f34bc356ef6dcf5eb03df9ccd94d8f1c))
+
+## [1.90.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.89.0...v1.90.0) (2025-05-15)
+
+### Features
+
+-   **three:** added new page sketches-page and new hook connectglslforwebgl ([cc4e849](https://github.com/DKMFzF/frontend-portfolio/commit/cc4e84960ac169fcba3b2fbd526b220887189c10))
+
 ## [1.89.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.88.0...v1.89.0) (2025-05-12)
 
 ### Features